fix decimal→int conversion
payment.amount (and refund.amount) are of type Decimal.

The correct way to convert to cents is to perform the multiplication with 100 in
the Decimal type and then convert to int:

  > int(Decimal('81.35') * 100)
  8135

Before this fix, the code inadvertantly stripped of one cent with certain
amounts, for example:

  > float(Decimal('81.35')) * 100
  8134.999999999999

Interestingly enough, the inverse operation float(cents) / 100 does seem to be
safe.

+++ pretix_datatrans/payment.py
@@ -46,6 +46,8 @@ class Datatrans(BasePaymentProvider):
         return True
 
     def execute_refund(self, refund: OrderRefund):
+        # refund docs:
+        # https://docs.pretix.eu/en/latest/development/implementation/models.html#pretix.base.models.OrderRefund
         payment_info = refund.payment.info_data
         if not payment_info:
             raise PaymentException(_("datatrans: no payment info found"))
@@ -68,7 +70,8 @@ class Datatrans(BasePaymentProvider):
             json={
                 "currency": self.event.currency,
                 "refno": refund.full_id,
-                "amount": float(refund.amount) * 100,
+                # refund.amount is of type decimal.Decimal
+                "amount": int(refund.amount * 100),
             },
             auth=HTTPBasicAuth(
                 gs.settings.payment_datatrans_merchant_id,
@@ -84,6 +87,8 @@ class Datatrans(BasePaymentProvider):
         logger.info("datatrans credit body = %s" % body)
 
     def execute_payment(self, request: HttpRequest, payment: OrderPayment):
+        # OrderPayment docs:
+        # https://docs.pretix.eu/en/latest/development/implementation/models.html#pretix.base.models.OrderPayment
         gs = GlobalSettingsObject()
         # initialize transaction by calling datatrans API
         transactions_url = "https://api.sandbox.datatrans.com/v1/transactions"
@@ -119,7 +124,8 @@ class Datatrans(BasePaymentProvider):
             json={
                 "currency": self.event.currency,
                 "refno": payment.full_id,
-                "amount": float(payment.amount) * 100,
+                # payment.amount is of type decimal.Decimal
+                "amount": int(payment.amount * 100),
                 "paymentMethods": payment_methods,
                 "redirect": {
                     "successUrl": success_url,
